policeParser to generic

+++ iTechArt.Domain/ModelInterfaces/IDashboardInfo.cs
@@ -6,12 +6,12 @@
         /// Gets count of pupils
         /// </summary>
         
-        public int PupilCount { get; }
+        public long PupilCount { get; }
 
         /// <summary>
         /// Gets count of students
         /// </summary>
-        public int StudentCount { get; }
+        public long StudentCount { get; }
 
         /// <summary>
         /// Gets count of groceries
